sidebar: create library
