+++ command_parse.c
@@ -56,9 +56,11 @@
 #include "mx.h"
 #include "myvar.h"
 #include "options.h"
-#include "sidebar.h"
 #include "version.h"
 #include "imap/lib.h"
+#ifdef USE_SIDEBAR
+#include "sidebar/lib.h"
+#endif
 #ifdef ENABLE_NLS
 #include <libintl.h>
 #endif
@@ -1022,54 +1024,6 @@ enum CommandResult parse_my_hdr(struct Buffer *buf, struct Buffer *s,
 
   return MUTT_CMD_SUCCESS;
 }
-
-#ifdef USE_SIDEBAR
-/**
- * parse_path_list - Parse the 'sidebar_whitelist' command - Implements Command::parse()
- */
-enum CommandResult parse_path_list(struct Buffer *buf, struct Buffer *s,
-                                   intptr_t data, struct Buffer *err)
-{
-  struct Buffer *path = mutt_buffer_pool_get();
-
-  do
-  {
-    mutt_extract_token(path, s, MUTT_TOKEN_BACKTICK_VARS);
-    mutt_buffer_expand_path(path);
-    add_to_stailq((struct ListHead *) data, mutt_b2s(path));
-  } while (MoreArgs(s));
-  mutt_buffer_pool_release(&path);
-
-  return MUTT_CMD_SUCCESS;
-}
-#endif
-
-#ifdef USE_SIDEBAR
-/**
- * parse_path_unlist - Parse the 'unsidebar_whitelist' command - Implements Command::parse()
- */
-enum CommandResult parse_path_unlist(struct Buffer *buf, struct Buffer *s,
-                                     intptr_t data, struct Buffer *err)
-{
-  struct Buffer *path = mutt_buffer_pool_get();
-
-  do
-  {
-    mutt_extract_token(path, s, MUTT_TOKEN_BACKTICK_VARS);
-    /* Check for deletion of entire list */
-    if (mutt_str_equal(mutt_b2s(path), "*"))
-    {
-      mutt_list_free((struct ListHead *) data);
-      break;
-    }
-    mutt_buffer_expand_path(path);
-    remove_from_stailq((struct ListHead *) data, mutt_b2s(path));
-  } while (MoreArgs(s));
-  mutt_buffer_pool_release(&path);
-
-  return MUTT_CMD_SUCCESS;
-}
-#endif
 
 /**
  * parse_set - Parse the 'set' family of commands - Implements Command::parse()
