+++ menu.c
@@ -49,7 +49,7 @@
 #include "pattern.h"
 #include "protos.h"
 #ifdef USE_SIDEBAR
-#include "sidebar.h"
+#include "sidebar/lib.h"
 #endif
 
 /* These Config Variables are only used in menu.c */
