+++ mutt_commands.c
@@ -42,6 +42,9 @@
 #include "mutt_globals.h"
 #include "mutt_lua.h"
 #include "score.h"
+#ifdef USE_SIDEBAR
+#include "sidebar/lib.h"
+#endif
 
 // clang-format off
 const struct Command Commands[] = {
@@ -108,7 +111,7 @@ const struct Command Commands[] = {
   { "setenv",              parse_setenv,           MUTT_SET_SET },
   { "shutdown-hook",       mutt_parse_hook,        MUTT_SHUTDOWN_HOOK | MUTT_GLOBAL_HOOK },
 #ifdef USE_SIDEBAR
-  { "sidebar_whitelist",   parse_path_list,        IP &SidebarWhitelist },
+  { "sidebar_whitelist",   sb_parse_whitelist,     0 },
 #endif
   { "source",              parse_source,           0 },
   { "spam",                parse_spam_list,        MUTT_SPAM },
@@ -146,7 +149,7 @@ const struct Command Commands[] = {
   { "unset",               parse_set,              MUTT_SET_UNSET },
   { "unsetenv",            parse_setenv,           MUTT_SET_UNSET },
 #ifdef USE_SIDEBAR
-  { "unsidebar_whitelist", parse_path_unlist,      IP &SidebarWhitelist },
+  { "unsidebar_whitelist", sb_parse_unwhitelist,   0 },
 #endif
   { "unsubjectrx",         parse_unsubjectrx_list, 0 },
   { "unsubscribe",         parse_unsubscribe,      0 },
