+++ sidebar/lib.h
@@ -5,7 +5,7 @@
  * @authors
  * Copyright (C) 2004 Justin Hibbits <jrh29@po.cwru.edu>
  * Copyright (C) 2004 Thomer M. Gil <mutt@thomer.com>
- * Copyright (C) 2015-2016 Richard Russon <rich@flatcap.org>
+ * Copyright (C) 2015-2020 Richard Russon <rich@flatcap.org>
  *
  * @copyright
  * This program is free software: you can redistribute it and/or modify it under
@@ -22,14 +22,16 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef MUTT_SIDEBAR_H
-#define MUTT_SIDEBAR_H
+#ifndef MUTT_SIDEBAR_LIB_H
+#define MUTT_SIDEBAR_LIB_H
 
 #include <stdbool.h>
+#include <stdint.h>
+#include "mutt/lib.h"
+#include "mutt_commands.h"
 
 struct Mailbox;
 struct MuttWindow;
-struct NotifyCallback;
 
 /* These Config Variables are only used in sidebar.c */
 extern short C_SidebarComponentDepth;
@@ -47,6 +49,7 @@ extern short C_SidebarSortMethod;
 extern bool  C_SidebarVisible;
 extern short C_SidebarWidth;
 
+extern struct ListHead SidebarWhitelist;
 
 /**
  * enum SidebarNotification - what happened to a mailbox
@@ -59,10 +62,14 @@ enum SidebarNotification
 };
 
 void            sb_change_mailbox  (int op);
+bool            select_next        (void);
 void            sb_draw            (struct MuttWindow *win);
 struct Mailbox *sb_get_highlight   (void);
 void            sb_notify_mailbox  (struct Mailbox *m, enum SidebarNotification sbn);
 int             sb_observer        (struct NotifyCallback *nc);
 void            sb_set_open_mailbox(struct Mailbox *m);
 
-#endif /* MUTT_SIDEBAR_H */
+enum CommandResult sb_parse_unwhitelist(struct Buffer *buf, struct Buffer *s, intptr_t data, struct Buffer *err);
+enum CommandResult sb_parse_whitelist  (struct Buffer *buf, struct Buffer *s, intptr_t data, struct Buffer *err);
+
+#endif /* MUTT_SIDEBAR_LIB_H */
